requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.26.146
botocore==1.29.146
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
dj-database-url==2.0.0
Django==4.2.1
django-active-link==0.1.8
django-autoslug==1.9.9
django-bootstrap4==23.1
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-countries==7.5.1
django-credit-cards==0.4.1
django-crispy-forms==1.14.0
django-filter==23.2
django-js-asset==2.0.0
django-livereload-server==0.4
django-mathfilters==1.0.0
django-phone-field==1.8.1
django-social-share==2.3.0
django-storages==1.13.2
django-template-maths==0.2.0
django-tinymce==3.6.1
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Pillow==10.1.0
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.4.0
